Fixes #1 Thumbnails in embeds are not loading
Since Discord seems to sometimes block embedded images from other CDNs,
such as from Twitch, instead download the thumbnail artwork to a temporary
file which is attached to the embed and sent. The thumbnail will then be hosted
on the Discord CDN and it should then be shown properly in stream shills.